+++ src/pages/Blog/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import { LayoutContainer } from '../../styles/layout';
+import { LayoutContainer } from '../../styles/layouts';
 
 export const Blog = () => {
   return <LayoutContainer></LayoutContainer>;
