+++ src/pages/FaqDetail/index.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
-import { BannerWrapper, ContainerInner, TopMargin } from '../../Layout';
+import {
+  BannerWrapper,
+  ContainerInner,
+  LayoutContainer,
+  TopMargin,
+} from '../../styles/layouts';
 import { Banner } from '../../img/Banner';
 import YellowBanner from '../../img/YellowBanner.png';
-import { LayoutContainer } from '../../styles/layout';
+
 import { Title } from '../../components/common/Title/title';
 import { Faq } from '../../api/faq';
 import { useNavigate, useParams } from 'react-router-dom';
@@ -20,17 +25,15 @@ const FaqDetail = () => {
       </BannerWrapper>
       <TopMargin />
       <LayoutContainer>
-        <ContainerInner>
-          <BackToFaq onClick={() => navigate('/faq')}>← 질문 목록</BackToFaq>
-          <QuestionTitle>{faqData?.question}</QuestionTitle>
-          <TopMargin />
-          <AnswerWrapper>
-            {faqData?.answer.split('\n').map((text, id) => (
-              <AnswerText key={id}>{text}</AnswerText>
-            ))}
-          </AnswerWrapper>
-          <TopMargin />
-        </ContainerInner>
+        <BackToFaq onClick={() => navigate('/faq')}>← 질문 목록</BackToFaq>
+        <QuestionTitle>{faqData?.question}</QuestionTitle>
+        <TopMargin />
+        <AnswerWrapper>
+          {faqData?.answer.split('\n').map((text, id) => (
+            <AnswerText key={id}>{text}</AnswerText>
+          ))}
+        </AnswerWrapper>
+        <TopMargin />
       </LayoutContainer>
     </>
   );
