+++ src/pages/index.tsx
@@ -33,7 +33,7 @@ export const Pages = () => {
         <Route path={'/signin'} element={<SignIn />} />
         <Route path={'/signup'} element={<SignUp />} />
         <Route path={'/apply'} element={<Apply />} />
-        <Route path={'/admin'} element={<Admin />} />
+        <Route path={'/admin/*'} element={<Admin />} />
         <Route path={'/onboarding/*'} element={<OnBoarding />} />
       </>
     </Routes>
