--- src/styles/layout.ts
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-
-export const LayoutContainer = styled.div`
-  position: relative;
-  width: 100%;
-  max-width: 1140px;
-  margin: 0 auto;
-  min-width: 320px;
-`;
